Updated wall following (frame for the actual code)
Robotathon project was removed because the uvproj target and other
configs are difficult to set in a brand new project so we are keeping
the RASdemo uvproj  for that simple fact. All of the RASdemo files
should be removed from the uvproj (but are still in the folder) and
robotathon.c contains the main method for the (now wall following but
soon to be full event) code.